READY TO DEPLOY

+++ src/features/category/components/data-table/ColumnsCategories.tsx
@@ -5,12 +5,23 @@ import { Button } from "@/components/ui/button";
 import {
 	DropdownMenu,
 	DropdownMenuContent,
-	DropdownMenuItem,
 	DropdownMenuLabel,
 	DropdownMenuSeparator,
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { useDeleteCategory } from "../../index";
+import {
+	AlertDialog,
+	AlertDialogAction,
+	AlertDialogCancel,
+	AlertDialogContent,
+	AlertDialogDescription,
+	AlertDialogFooter,
+	AlertDialogHeader,
+	AlertDialogTitle,
+	AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import { MainButton } from "@/components/MainButton";
 
 type TColumnsCategories = {
 	id: number;
@@ -42,14 +53,36 @@ export const ColumnsCategories: ColumnDef<TColumnsCategories>[] = [
 						</Button>
 					</DropdownMenuTrigger>
 					<DropdownMenuContent align="end">
-						<DropdownMenuLabel>Actions</DropdownMenuLabel>
+						<DropdownMenuLabel>Options</DropdownMenuLabel>
 						<DropdownMenuSeparator />
-						<DropdownMenuItem
-							onClick={() => deleteCategory(category.id)}
-							className="bg-red-600 text-white hover:!bg-red-600 hover:!text-white cursor-pointer"
-						>
-							Delete category
-						</DropdownMenuItem>
+						<AlertDialog>
+							<AlertDialogTrigger className="bg-red-600 duration-200 text-white hover:!bg-red-800 hover:!text-white cursor-pointer px-2 py-1 flex rounded-sm">
+								Delete category
+							</AlertDialogTrigger>
+							<AlertDialogContent className="bg-black text-white">
+								<AlertDialogHeader>
+									<AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+									<AlertDialogDescription>
+										This action cannot be undone. This will permanently delete
+										your account and remove your data from our servers.
+									</AlertDialogDescription>
+								</AlertDialogHeader>
+								<AlertDialogFooter>
+									<AlertDialogCancel className="bg-black text-white duration-200">
+										Cancel
+									</AlertDialogCancel>
+									<AlertDialogAction
+										asChild
+										onClick={() => deleteCategory(category.id)}
+									>
+										<MainButton
+											className="!bg-golden duration-200 hover:!bg-white hover:text-black"
+											title="Continue"
+										/>
+									</AlertDialogAction>
+								</AlertDialogFooter>
+							</AlertDialogContent>
+						</AlertDialog>
 					</DropdownMenuContent>
 				</DropdownMenu>
 			);
